Code refactor, pagination, search criteria after favorites, changing a user's subscription

+++ services/contacts.js
@@ -1,7 +1,26 @@
 const { ObjectId } = require("mongoose").Types;
 const Contact = require("../models/contact");
 
-const getAllContacts = async () => await Contact.find({}).lean();
+const getAllContacts = async (id, queries) => {
+  try {
+    const { page = 1, limit = 20, favorite = null } = queries;
+
+    if (!favorite) {
+      return await Contact.find({ owner: id })
+        .limit(+limit)
+        .skip((page - 1) * limit);
+    }
+
+    return await Contact.find({
+      owner: id,
+      favorite,
+    })
+      .limit(+limit)
+      .skip((page - 1) * limit);
+  } catch (err) {
+    return null;
+  }
+};
 
 const getSingleContact = async (contactId) => {
   let objectIdContact;
@@ -14,17 +33,23 @@ const getSingleContact = async (contactId) => {
   return await Contact.findOne({ _id: objectIdContact }).lean();
 };
 
-const createContact = async (body) => await Contact.create(body);
+const createContact = async (userId, body) => {
+  const contact = await Contact.findOne({ ...body, owner: userId });
+  if (contact) {
+    return null;
+  }
 
-const deleteContact = async (contactId) => {
+  return await Contact.create({ ...body, owner: userId });
+}
+
+const deleteContact = async (userId, contactId) => {
   let objectIdContact;
   try {
     objectIdContact = ObjectId(contactId);
   } catch (err) {
     return null;
   }
-  console.log(objectIdContact);
-  return await Contact.findOneAndDelete({ _id: objectIdContact});
+  return await Contact.findOneAndDelete({ _id: objectIdContact, owner: userId });
 };
 
 const updateContact = async (contactId, body) => {
@@ -35,17 +60,17 @@ const updateContact = async (contactId, body) => {
     return null;
   }
   return await Contact.findOneAndUpdate(
-    { _id:  objectIdContact },
+    { _id: objectIdContact },
     {
       $set: body,
     },
     {
-			new: true,
-			runValidators: true,
-			strict: "throw",
-		}
-  )
-}
+      new: true,
+      runValidators: true,
+      strict: "throw",
+    }
+  );
+};
 
 module.exports = {
   getAllContacts,
